design: 피드 이모지 스크롤에서 wrap으로 변경
